+++ lib/active_record/connection_adapters/postgresql_proxy_adapter.rb
@@ -41,3 +41,6 @@ if ActiveRecordProxyAdapters::ActiveRecordContext.active_record_v7_2_or_greater?
     "active_record/connection_adapters/postgresql_proxy_adapter"
   )
 end
+
+ActiveSupport.run_load_hooks(:active_record_postgresqlproxyadapter,
+                             ActiveRecord::ConnectionAdapters::PostgreSQLProxyAdapter)